Refactor Syncer and Upsyncer VW common parts
