+++ pkg/virtual/syncer/upsyncer/transformer.go
@@ -32,8 +32,13 @@ import (
 	syncercontext "github.com/kcp-dev/kcp/pkg/virtual/syncer/context"
 )
 
-const UpsyncDiffAnnotationPrefix = "diff.upsync.workload.kcp.dev/"
+// UpsyncDiffAnnotationPrefix is an internal annotation used on downstream resources to specify a transformation
+// that should be applied during the Upsyncing of the resource to upstream.
+// Format of the annotation is JSONPatch
+const UpsyncDiffAnnotationPrefix = "internal.workload.kcp.dev/upsyncdiff"
 
+// UpsyncerResourceTransformer defines a very simple transformer which transforms the resource by applying a
+// the JSON patch found in the `internal.workload.kcp.dev/upsyncdiff` annotation
 type UpsyncerResourceTransformer struct{}
 
 func (rt *UpsyncerResourceTransformer) AfterRead(client dynamic.ResourceInterface, ctx context.Context, gvr schema.GroupVersionResource, upstreamResource *unstructured.Unstructured, eventType *watch.EventType, subresources ...string) (*unstructured.Unstructured, error) {
